CLOE-main/requirements.txt
torch
pyod
numpy
scikit-learn
umap-learn
matplotlib
argparse
torchvision
optuna
torchkde
# adbench